Move to separate file to avoid cyclic imports
