+++ qcodes/tests/dataset/test_guid_helpers.py
@@ -1,0 +1,58 @@
+from collections import defaultdict
+from pathlib import Path
+import pytest
+
+import numpy as np
+
+from qcodes.dataset.experiment_container import new_experiment
+from qcodes.dataset.measurements import Measurement
+from qcodes.dataset.sqlite.database import initialised_database_at
+from qcodes.instrument.parameter import Parameter
+
+
+from qcodes.dataset.guid_helpers import guids_from_dir, guids_from_list_str
+
+
+def test_guids_from_dir(tmp_path: Path) -> None:
+    def generate_local_run(dbpath: Path) -> str:
+        with initialised_database_at(str(dbpath)):
+            new_experiment(sample_name="fivehundredtest_sample",
+                           name="fivehundredtest_name")
+
+            p1 = Parameter('Voltage', set_cmd=None)
+            p2 = Parameter('Current', get_cmd=lambda: np.random.randn())
+
+            meas = Measurement()
+            meas.register_parameter(p1).register_parameter(p2, setpoints=[p1])
+
+            with meas.run() as datasaver:
+                for v in np.linspace(0, 2, 250):
+                    p1(v)
+                    datasaver.add_result((p1, p1()), (p2, p2()))
+            guid = datasaver.dataset.guid
+            datasaver.flush_data_to_database(block=True)
+        return guid
+
+    paths_counts = [
+        (tmp_path / 'subdir1' / 'dbfile1.db', 2),
+        (tmp_path / 'subdir1' / 'dbfile2.db', 4),
+        (tmp_path / 'subdir2' / 'dbfile1.db', 1),
+        (tmp_path / 'dbfile1.db', 3),
+    ]
+    guids = defaultdict(list)
+    for path, count in paths_counts:
+        path.parent.mkdir(exist_ok=True, parents=True)
+        for _ in range(count):
+            guids[path].append(generate_local_run(path))
+    dbdict, guidsdict = guids_from_dir(tmp_path)
+    assert dbdict == guids
+
+
+def test_guids_from_list_str() -> None:
+    guids = ['07fd7195-c51e-44d6-a085-fa8274cf00d6',
+             '070d7195-c51e-44d6-a085-fa8274cf00d6']
+    assert guids_from_list_str('') == tuple()
+    assert guids_from_list_str(str(guids)) == tuple(guids)
+    assert guids_from_list_str(str([guids[0]])) == (guids[0],)
+    assert guids_from_list_str(str(tuple(guids))) == tuple(guids)
+    assert guids_from_list_str(str(guids[0])) == (guids[0],)
